requirements.txt
jira==3.2.0
PyYAML==6.0
memoization==0.4.0